Add pattern state on ant behavior
